added logic for player movement

+++ Core/Entities/Map/Tiles/DesertTile.cs
@@ -2,7 +2,7 @@
 {
 	public class DesertTile : MapTileBase
 	{
-		public override double MovesCostToMoveOnTile => 4;
+		public override uint MovesCostToMoveOnTile => 4;
 		public override string Description => "Desert";
 		public override TerrainTypes TerrainType => TerrainTypes.DESERT;
 	}
