+++ Core/Entities/Map/Tiles/ForestTile.cs
@@ -2,7 +2,7 @@
 {
 	public class ForestTile : MapTileBase
 	{
-		public override double MovesCostToMoveOnTile => 2;
+		public override uint MovesCostToMoveOnTile => 2;
 		public override string Description => "Forest";
 		public override TerrainTypes TerrainType => TerrainTypes.FOREST;
 	}
