+++ Core/Entities/Map/Tiles/GrassTile.cs
@@ -2,7 +2,7 @@
 {
 	public class GrassTile : MapTileBase
 	{
-		public override double MovesCostToMoveOnTile => 1;
+		public override uint MovesCostToMoveOnTile => 1;
 		public override string Description => "Grass";
 		public override TerrainTypes TerrainType => TerrainTypes.GRASS;
 	}
