+++ Core/Entities/Map/Tiles/MapTileBase.cs
@@ -2,7 +2,7 @@
 {
 	public abstract class MapTileBase
 	{
-		public abstract double MovesCostToMoveOnTile { get; }
+		public abstract uint MovesCostToMoveOnTile { get; }
 		public abstract string Description { get; }
 		public abstract TerrainTypes TerrainType { get; }
