+++ Core/Entities/Map/Tiles/WaterTile.cs
@@ -2,7 +2,7 @@
 {
 	public class WaterTile : MapTileBase
 	{
-		public override double MovesCostToMoveOnTile => Defines.Map.IMPASSABLE_FOR_NONFLYING_CREATURES_TERRAIN_MOVEMENT_COST;
+		public override uint MovesCostToMoveOnTile => Defines.Map.IMPASSABLE_FOR_NONFLYING_CREATURES_TERRAIN_MOVEMENT_COST;
 		public override string Description => "Water";
 		public override TerrainTypes TerrainType => TerrainTypes.WATER;
 	}
